Changed the floats to doubles although i disagree that it's necessary, and set imagery-decimation and imager-terrain pairs to use options.setExclusive()

+++ earth_enterprise/src/fusion/rasterfuse/TmeshGenerator.cpp
@@ -58,7 +58,7 @@ TmeshGenerator::Generate(float *srcSamples,
                          etArray<uchar> &compressed,
                          const size_t reserve_size,
                          bool decimate,
-                         float decimation_threshold
+                         double decimation_threshold
                         )
 {
   // reset my temporaries - they live in the class so I don't have to
