+++ earth_enterprise/src/fusion/rasterfuse/TmeshPackgenTraverser.h
@@ -76,13 +76,13 @@ class TmeshWorkItem {
   uint numPiecesUsed;
   TmeshGenerator generator;
   bool decimate_;
-  float decimation_threshold_;
+  double decimation_threshold_;
 
   TmeshWorkItem(PacketFileWriter &_writer,
                 const AttributionByExtents &_attributions,
                 const khTilespaceFlat &_sampleTilespace,
                 bool decimate,
-                float decimation_threshold);
+                double decimation_threshold);
 
   void DoWork(TmeshPrepItem *prep);
   void DoWrite(khMTProgressMeter *progress);
@@ -107,7 +107,7 @@ class TmeshPackgenTraverser : public PackgenTraverser<TmeshPackgenBaseConfig> {
   TmeshPackgenTraverser(const PacketLevelConfig &config,
                         geFilePool &file_pool_,
                         const std::string &output,
-                        float decimation_threshold
+                        double decimation_threshold
                        );
 
   virtual TmeshPrepItem* NewPrepItem(void);
